+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueUpdate.kt
@@ -141,10 +141,7 @@ class OverdueUpdate(
     return if (canGeneratePdf)
       OpenSelectDownloadFormatDialog(selectedAppointmentIds)
     else
-      ScheduleDownload(
-          fileFormat = CSV,
-          selectedAppointmentIds = selectedAppointmentIds
-      )
+      ScheduleDownload(fileFormat = CSV)
   }
 
   private fun openDialogForShareEffect(selectedAppointmentIds: Set<UUID>): OverdueEffect {