+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Implement `OverdueAppointmentSelector`
 - Use overdue appointment selector in overdue screen
 - Use overdue appointment selector in overdue search screen
+- Use overdue appointment selector to fetch selected ids in overdue download worker
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-11-8325